refactor: 확장 함수 파일 분리
