+++ app/src/main/java/woowacourse/movie/utils/extensions/ContextExt.kt
@@ -1,0 +1,8 @@
+package woowacourse.movie.utils
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.showToast(message: String) {
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
