+++ app/src/main/java/woowacourse/movie/utils/extensions/FragmentManagerExt.kt
@@ -1,0 +1,8 @@
+package woowacourse.movie.utils
+
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+
+fun FragmentManager.commit(content: FragmentTransaction.() -> Unit) {
+    beginTransaction().apply(content).commit()
+}
